evaluator-gcp-cloudsql/scripts/requirements.txt
google-cloud-sql
google-cloud-sql-connector
google-cloud-resource-manager
mysql-connector-python
